refactor: moved html pkg to login subpackage
